config: Move envfile tests to new file
We want to add config related tests to config_test.go. Move envfile
related tests to their own file.
